+++ pkg/providers/property/provider.go
@@ -120,6 +120,7 @@ func (p *Subprovider) FrameworkDataSources() []func() datasource.DataSource {
 		NewHostnameAuditHistoryDataSource,
 		NewHostnamesDiffDataSource,
 		NewIncludeDataSource,
+		NewReportingGroupDataSource,
 	}
 }